requirements.txt
torch
torchvision
pandas
matplotlib
scikit-learn
fastapi
requests
pillow
uvicorn